Level Design And Boss Fight

+++ Assets/Scriptable Objects/FloatVariable.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName ="Float Variable",menuName ="Data/Float Variable")]
 public class FloatVariable : ScriptableObject
 {
-    [SerializeField] public float value;
+    [SerializeField] public int value;
 }
